+++ CHANGELOG.md
@@ -21,6 +21,7 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 - Inlined `Decoder` into `SchemaValidator` ([#157]), by [@localheinz]
 - Disallowed injection of `Validator` into `SchemaValidator` ([#158]), by [@localheinz]
 - Removed `Schema` ([#161]), by [@localheinz]
+- Composed `Error` into `Result` ([#166]), by [@localheinz]
 
 ## [`1.0.0`][1.0.0]
 
@@ -52,5 +53,6 @@ For a full diff see [`dcd4cfb...1.0.0`][dcd4cfb...1.0.0].
 [#163]: https://github.com/ergebnis/json-schema-validator/pull/163
 [#164]: https://github.com/ergebnis/json-schema-validator/pull/164
 [#165]: https://github.com/ergebnis/json-schema-validator/pull/165
+[#166]: https://github.com/ergebnis/json-schema-validator/pull/166
 
 [@localheinz]: https://github.com/localheinz